refactor(types): Move schemas
