+++ libs/types/src/lib/index.ts
@@ -1,5 +1,2 @@
-export * from './participant.schema';
-export * from './poll.schema';
-export * from './poll-event.schema';
-export * from './settings';
+export * from './schema';
 export * from './dto';
